CR-Mall/requirements.txt
alembic==1.0.2
aniso8601==4.0.1
awscli==1.16.56
blinker==1.4
boto3==1.9.44
botocore==1.12.46
certifi==2018.10.15
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.3.9
connexion==2.0.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.13.1
Flask-Mail==0.9.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.15.0
mysqlclient==1.3.13
openapi-spec-validator==0.2.4
pkg-resources==0.0.0
pyasn1==0.4.4
PyJWT==1.6.4
pyotp==2.2.7
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyYAML==3.13
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.24.1
Werkzeug==0.14.1
